Updated ev_data_context to inherit from a common Limits base class
