+++ iso15118/secc/states/iso15118_20_states.py
@@ -1315,7 +1315,7 @@ class ACChargeParameterDiscovery(StateSECC):
             ac_cpd_req.ac_params
         ):
             self.comm_session.evse_controller.ev_data_context.ev_rated_limits.update(
-                ac_cpd_req.ac_params
+                ac_cpd_req.ac_params.dict()
             )
             ac_params = (
                 await self.comm_session.evse_controller.get_ac_charge_params_v20(
@@ -1326,7 +1326,7 @@ class ACChargeParameterDiscovery(StateSECC):
             ac_cpd_req.bpt_ac_params
         ):
             self.comm_session.evse_controller.ev_data_context.ev_rated_limits.update(
-                ac_cpd_req.bpt_ac_params
+                ac_cpd_req.bpt_ac_params.dict()
             )
             bpt_ac_params = (
                 await self.comm_session.evse_controller.get_ac_charge_params_v20(
@@ -1416,14 +1416,14 @@ class ACChargeLoop(StateSECC):
         if selected_energy_service.service == ServiceV20.AC:
             if control_mode == ControlMode.SCHEDULED:
                 self.comm_session.evse_controller.ev_data_context.ev_session_context.update(  # noqa
-                    ac_charge_loop_req.scheduled_params
+                    ac_charge_loop_req.scheduled_params.dict()
                 )
                 scheduled_params = await self.comm_session.evse_controller.get_ac_charge_loop_params_v20(  # noqa
                     ControlMode.SCHEDULED, ServiceV20.AC
                 )
             elif control_mode == ControlMode.DYNAMIC:
                 self.comm_session.evse_controller.ev_data_context.ev_session_context.update(  # noqa
-                    ac_charge_loop_req.dynamic_params
+                    ac_charge_loop_req.dynamic_params.dict()
                 )
                 dynamic_params = await self.comm_session.evse_controller.get_ac_charge_loop_params_v20(  # noqa
                     ControlMode.DYNAMIC, ServiceV20.AC
@@ -1431,14 +1431,14 @@ class ACChargeLoop(StateSECC):
         elif selected_energy_service.service == ServiceV20.AC_BPT:
             if control_mode == ControlMode.SCHEDULED:
                 self.comm_session.evse_controller.ev_data_context.ev_session_context.update(  # noqa
-                    ac_charge_loop_req.bpt_scheduled_params
+                    ac_charge_loop_req.bpt_scheduled_params.dict()
                 )
                 bpt_scheduled_params = await self.comm_session.evse_controller.get_ac_charge_loop_params_v20(  # noqa
                     ControlMode.SCHEDULED, ServiceV20.AC_BPT
                 )  # noqa
             else:
                 self.comm_session.evse_controller.ev_data_context.ev_session_context.update(  # noqa
-                    ac_charge_loop_req.bpt_dynamic_params
+                    ac_charge_loop_req.bpt_dynamic_params.dict()
                 )
                 bpt_dynamic_params = await self.comm_session.evse_controller.get_ac_charge_loop_params_v20(  # noqa
                     ControlMode.DYNAMIC, ServiceV20.AC_BPT
@@ -1537,7 +1537,7 @@ class DCChargeParameterDiscovery(StateSECC):
             dc_cpd_req.dc_params
         ):
             self.comm_session.evse_controller.ev_data_context.ev_rated_limits.update(
-                dc_cpd_req.dc_params
+                dc_cpd_req.dc_params.dict()
             )
             dc_params = (
                 await self.comm_session.evse_controller.get_dc_charge_params_v20(
@@ -1548,7 +1548,7 @@ class DCChargeParameterDiscovery(StateSECC):
             dc_cpd_req.bpt_dc_params
         ):
             self.comm_session.evse_controller.ev_data_context.ev_rated_limits.update(
-                dc_cpd_req.bpt_dc_params
+                dc_cpd_req.bpt_dc_params.dict()
             )
             bpt_dc_params = (
                 await self.comm_session.evse_controller.get_dc_charge_params_v20(
@@ -1804,20 +1804,20 @@ class DCChargeLoop(StateSECC):
         if selected_energy_service.service == ServiceV20.DC:
             if control_mode == ControlMode.SCHEDULED:
                 self.comm_session.evse_controller.ev_data_context.ev_session_context.update(  # noqa
-                    dc_charge_loop_req.scheduled_params
+                    dc_charge_loop_req.scheduled_params.dict()
                 )
             elif control_mode == ControlMode.DYNAMIC:
                 self.comm_session.evse_controller.ev_data_context.ev_session_context.update(  # noqa
-                    dc_charge_loop_req.dynamic_params
+                    dc_charge_loop_req.dynamic_params.dict()
                 )
         elif selected_energy_service.service == ServiceV20.DC_BPT:
             if control_mode == ControlMode.SCHEDULED:
                 self.comm_session.evse_controller.ev_data_context.ev_session_context.update(  # noqa
-                    dc_charge_loop_req.bpt_scheduled_params
+                    dc_charge_loop_req.bpt_scheduled_params.dict()
                 )
             else:
                 self.comm_session.evse_controller.ev_data_context.ev_session_context.update(  # noqa
-                    dc_charge_loop_req.bpt_dynamic_params
+                    dc_charge_loop_req.bpt_dynamic_params.dict()
                 )
         else:
             logger.error(
